requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==43.0.0
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
Faker==26.0.0
filelock==3.13.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
openai==1.31.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
SQLAlchemy-serializer==1.4.12
stack-data==0.6.3
tomli==2.0.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.19.2
